meal rest controller add but not all

+++ src/main/java/ru/javawebinar/topjava/web/meal/MealRestController.java
@@ -7,11 +7,11 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.service.MealService;
 import ru.javawebinar.topjava.to.MealWithExceed;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Controller
@@ -21,7 +21,11 @@ public class MealRestController {
     private MealService service;
 
     public Meal get(int id) {
-        return service.get(id, AuthorizedUser.id());
+        int userId = AuthorizedUser.id();
+        if (null == service.get(id, userId)) {
+            throw new NotFoundException("meal with id " + id + " not found");
+        }
+        return service.get(id, userId);
     }
 
     public Meal create(Meal meal) {
@@ -44,10 +48,11 @@ public class MealRestController {
     public List<MealWithExceed> getBeetwen(LocalDate startDate, LocalTime startTime, LocalDate endDate, LocalTime endTime) {
         int userId = AuthorizedUser.id();
         return MealsUtil.getWithExceeded(
-                service.getBetweenDates(startDate, endDate, userId)
+                service.getBetweenDates(startDate == null ? LocalDate.MIN : startDate,
+                                        endDate == null ? LocalDate.MAX: endDate, userId)
                         .stream()
-                        .filter(meal -> meal.getDateTime().toLocalTime().isAfter(startTime)
-                                     && meal.getDateTime().toLocalTime().isBefore(endTime))
+                        .filter(meal -> meal.getDateTime().toLocalTime().isAfter(startTime == null ? LocalTime.MIN : startTime)
+                                     && meal.getDateTime().toLocalTime().isBefore(endTime == null ? LocalTime.MAX : endTime))
                         .collect(Collectors.toList()),
                                  AuthorizedUser.getCaloriesPerDay()
         );
@@ -55,13 +60,17 @@ public class MealRestController {
 
     public void delete(int id) {
         int userId = AuthorizedUser.id();
-        Objects.requireNonNull(service.get(id, userId), "meal don't exist");
+        if (null == service.get(id, userId)) {
+            throw new NotFoundException("meal with id " + id + " not found");
+        }
         service.delete(id, userId);
     }
 
     public void update(Meal meal, int id) {
         int userId = AuthorizedUser.id();
-        Objects.requireNonNull(service.get(id, userId), "meal don't exist");
+        if (null == service.get(id, userId)) {
+            throw new NotFoundException("meal with id " + id + " not found");
+        }
         service.update(meal, userId);
     }
